int to long

+++ src/main/java/via/sdj3/slaughterhouse/service/AnimalService.java
@@ -4,5 +4,5 @@ import via.sdj3.slaughterhouse.model.Animal;
 
 public interface AnimalService
 {
-    Animal registerAnimal( double weight, int registrationNumber, String origin);
+    Animal registerAnimal( double weight, long registrationNumber, String origin);
 }
